Chore: Defer version bumps to release script
Now extension-browser should be released as part of running the
global release script for the monorepo to ensure it stays in sync with
published documentation.
