+++ packages/extension-browser/package.json
@@ -100,7 +100,6 @@
     "node": ">=8.0.0"
   },
   "files": [
-    "dist/src",
     "images"
   ],
   "homepage": "https://webhint.io/",
@@ -112,7 +111,6 @@
     "webhint"
   ],
   "license": "Apache-2.0",
-  "main": "./dist/src/connector.js",
   "name": "@hint/extension-browser",
   "repository": {
     "directory": "packages/extension-browser",
@@ -121,13 +119,12 @@
   },
   "scripts": {
     "build": "npm-run-all build:* && webpack",
-    "build-release": "npm run clean && npm run bump-version && npm run build-release-packages",
+    "build-release": "npm run clean && npm run build-release-packages",
     "build-release-packages": "npm run build:assets && tsc --inlineSourceMap false --removeComments true && npm-run-all build-release:*",
     "build-release:chromium": "webpack --env.release --env.design=fluent && web-ext build -s ./dist/bundle -a ./dist/chromium --overwrite-dest && npm run lint-bundle-size ./dist/chromium",
     "build-release:firefox": "node scripts/drop-persistent-flag.js && webpack --env.release --env.design=photon && web-ext build -s ./dist/bundle -a ./dist/firefox --overwrite-dest && npm run lint-bundle-size ./dist/firefox",
     "build:assets": "cpx \"./{src,tests}/**/{!(*.ts),.!(ts),!(*.tsx),.!(tsx)}\" dist && npm-run-all prebuild:*",
     "build:ts": "tsc -b",
-    "bump-version": "node scripts/bump-version.js",
     "clean": "rimraf dist src/**/*.css.d.ts",
     "host-ui": "node scripts/host-ui.js",
     "lint": "npm-run-all lint:*",
@@ -139,7 +136,7 @@
     "prebuild:css": "tcm src",
     "prebuild:hints": "node ./scripts/import-hints.js",
     "prebuild:i18n": "node ./scripts/create-i18n.js && node ./scripts/merge-i18n.js",
-    "release": "npm run build-release && node scripts/submit-packages.js && npm publish",
+    "prebuild:version": "node ./scripts/copy-version.js",
     "stats": "webpack --json > stats.json",
     "stats-release": "webpack --env.release --json > release-stats.json",
     "test": "npm run lint && npm run build && npm run test-only && npm run build-release-packages",
